How about more refactoring?
